+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.2.30-2"></a>
+## [0.2.30-2](http://git.hyperchain.cn/ultramesh/flato-rbft/compare/v0.2.30-1...v0.2.30-2) (2021-04-09)
+
+
+### Performance Improvements
+
+* **rbft_impl.go:** batch add requests into txpool to improve the performance of look up bloom_filte ([0c367b2](http://git.hyperchain.cn/ultramesh/flato-rbft/commits/0c367b2)), closes [#flato-3340](http://git.hyperchain.cn/ultramesh/flato-rbft/issues/flato-3340)
+
+
+
+<a name="0.2.30-1"></a>
+## [0.2.30-1](http://git.hyperchain.cn/ultramesh/flato-rbft/compare/v0.2.30...v0.2.30-1) (2021-04-08)
+
+
+### Bug Fixes
+
+* save useful batches before state update to avoid infinite vc caused by missing request batch ([a06f89b](http://git.hyperchain.cn/ultramesh/flato-rbft/commits/a06f89b)), closes [#flato-3334](http://git.hyperchain.cn/ultramesh/flato-rbft/issues/flato-3334)
+
+
+
 <a name="0.2.30"></a>
 ## [0.2.30](http://git.hyperchain.cn/ultramesh/flato-rbft/compare/v0.2.29...v0.2.30) (2021-03-25)